requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.54.1
graphene==2.1.8
graphql-core==2.3.1
graphql-relay==2.0.1
h11==0.9.0
httptools==0.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
orjson==2.6.6
promise==2.3
pycryptodome==3.9.7
pydantic==1.5.1
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.23.0
Rx==1.6.1
six==1.14.0
starlette==0.13.2
ujson==2.0.3
urllib3==1.25.9
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
